// frontend/src/api/generated/models/bookingDto.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Backend API
 * A simple ASP.NET Core Web API template with best practices for building scalable and maintainable applications.
 * OpenAPI spec version: v1
 */

export interface BookingDto {
  id: number;
  roomId: number;
  roomName: string;
  startTime: string;
  endTime: string;
  bookedBy: string;
  /** @nullable */
  purpose?: string | null;
  createdAt: string;
}